% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iv.R
\name{ivBounds}
\alias{ivBounds}
\title{Instrumental Variables causal effects bounds

Get generalization of Balke and Pearl's IV bounds for model with \eqn{W -> X -> Y}.}
\usage{
ivBounds(p, epsilons = c(0, 1, 1, 1, 1, 1))
}
\arguments{
\item{p}{an array or vector of probabilities \eqn{P(X=x, Y=x, W=w)} or \eqn{P(X=x, Y=x | W=w)}, with \eqn{X} changing fastest, \eqn{W} slowest}

\item{epsilons}{vector of relaxation parameters (see details); defaults to Balke-Pearl bounds}
}
\description{
Instrumental Variables causal effects bounds

Get generalization of Balke and Pearl's IV bounds for model with \eqn{W -> X -> Y}.
}
\details{
Joint probabilities should be specified, but conditional probabilities are
sufficient (and permitted) if \code{epsilon[5] = epsilon[6] = 1}.  If either
doesn't hold and conditional probabilities are
supplied, then an error is returned.


\code{epsilons} is a vector of six positions corresponding to the relaxation parameters. In order:
       \enumerate{
       \item the maximum difference in the conditional probability of the outcome given everything else, as the instrument changes levels;
       \item the maximum difference in the conditional probability of the outcome given everything else, and the conditional distribution
           excluding latent variables for the instrument set at 0;
       \item the maximum difference in the conditional probability of the outcome given everything else, and the conditional distribution
           excluding latent variables for the instrument set at 1;
       \item the maximum difference in the conditional probability of the treatment given its causes, and the conditional distribution
           excluding latent variables
       \item the maximum ratio between the conditional distribution of the latent variable given the instrument and the
           marginal distribution of the latent variable. This has to be greater than or equal to 1;
       \item the minimum ratio between the conditional distribution of the latent variable given the instrument and the
           marginal distribution of the latent variable. This has to be in the interval (0, 1].
           }
           Setting this to \code{c(0,1,1,1,1,1)} corresponds to the ordinary IV model.
}

