% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/causal_class.R
\name{print.cfx}
\alias{print.cfx}
\title{Prints a CausalFX Problem Instance}
\usage{
\method{print}{cfx}(x, ...)
}
\arguments{
\item{x}{a \code{cfx} object.}

\item{...}{other parameters, ignored.}
}
\description{
Prints some of the information regarding a \code{\link{cfx}} object.
}
\details{
The information that is displayed includes the identifiers of the treatment and outcome variables,
the names of all variables and, if a theoretical causal graph is part of the object specification,
its causal structured described in terms of the parent ids for each variable.
}

