% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{print.summary.covsearch}
\alias{print.summary.covsearch}
\title{Print Summaries of Covariate Search Outputs}
\usage{
\method{print}{summary.covsearch}(x, ...)
}
\arguments{
\item{x}{an object of class "\code{summary.covsearch}", usually a result of a call to \code{\link{summary.covsearch}}.}

\item{...}{other parameters, ignored.}
}
\description{
Print output of \code{summary} method for class "\code{covsearch}".
}
\details{
Variable names with more than 20 characters are truncated when printing.
}

