% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iv.R
\name{print.summary.iv}
\alias{print.summary.iv}
\title{Print Summaries of Binary Instrumental Variable Analyses}
\usage{
\method{print}{summary.iv}(x, ...)
}
\arguments{
\item{x}{an object of class "\code{summary.iv}", usually a result of a call to \code{\link{summary.iv}}.}

\item{...}{other parameters, ignored.}
}
\description{
Print output of \code{summary} method for class "\code{iv}".
}
\details{
Variable names with more than 20 characters are truncated.
}

