% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wpp.R
\name{print.summary.wpp}
\alias{print.summary.wpp}
\title{Print Summaries of Witness Protection Program Outputs}
\usage{
\method{print}{summary.wpp}(x, ...)
}
\arguments{
\item{x}{an object of class "\code{summary.wpp}", usually a result of a call to \code{\link{summary.wpp}}.}

\item{...}{other parameters, ignored.}
}
\description{
Print output of \code{summary} method for class "\code{wpp}".
}
\details{
Variable names with more than 20 characters are truncated when printing.
}

