% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{summary.covsearch}
\alias{summary.covsearch}
\title{Summarize Covariate Search Outputs}
\usage{
\method{summary}{covsearch}(object, ...)
}
\arguments{
\item{object}{an object of class "\code{covsearch}", usually a result of a call to \code{\link{covsearch}}.}

\item{...}{other parameters, ignored.}
}
\value{
Besides fields inherented from the \code{covsearch} object, a list summary statistics is included:
 \item{\code{q}}{an array of 5 entries corresponding to evenly space quantiles of the ACE of the highest scoring
                 witness/admissible set pair.}
 \item{\code{ci}}{95\% marginal posterior credible interval for the chosen ACE.}
}
\description{
\code{summary} method for class "\code{covsearch}".
}
\seealso{
The model fitting function \code{\link{covsearch}}.
}

