% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iv.R
\name{summary.iv}
\alias{summary.iv}
\title{Summarize Binary Instrumental Variable Analyses}
\usage{
\method{summary}{iv}(object, ...)
}
\arguments{
\item{object}{an object of class "\code{iv}", usually a result of a call to \code{\link{iv}}.}

\item{...}{other parameters, ignored.}
}
\value{
Besides fields from the \code{iv} object, this includes a list summary statistics,
 \item{\code{lq}}{an array of 5 entries corresponding to evenly space quantiles of the lower bound ("minimum", 25\%, median, 75%, "maximum") as
                  given by a Monte Carlo approximation.}
 \item{\code{lci}}{95\% marginal posterior credible interval for lower bound.}
 \item{\code{uq}}{an array of 5 entries corresponding to evenly space quantiles of the upper bound ("minimum", 25\%, median, 75%, "maximum") as
                  given by a Monte Carlo approximation.}
 \item{\code{uci}}{95\% marginal posterior credible interval for upper bound.}
}
\description{
\code{summary} method for class "\code{iv}".
}
\seealso{
The model fitting function \code{\link{iv}}.
}

